+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.0.3
+
+- added `fakeResponse.generateMock.resourceTypeFilters` - to filter resourceType while generating mock.
+- removed `Fake Response: Restart Server`. instead use the same `Fake Response: Start Server` to refresh the server
+
 ## v0.0.2
 
 - bug fix. get updated injectors.js file on restarting server.